refaktoriserer AppProvider og lager EøsProvider
